Add files via upload
This is code deals with the statistical analysis of the World Bank dataset wdidata20.csv. It calculates the correlation between two variables within the wdi dataset. It also contains a linear regression to determine if the relationship between the variables is statically significant or not. The results of linear regression and correlation are visualized in the form of a scatter plot. The study itself investigates if there exists a relationship between a person's access to electricity and female primary school graduation rate. The findings of the study observe that the relationship is statistically significant. Therefore, the relationship between the independent variable  "access to electricity" and the dependent variable "female primary school graduation rate" is causal since a statistically significant relationship allows us to reject the null hypothesis, which claims that "access to electricity" is unrelated to "female primary school graduation rate". The results of the study are significant on a substantial level because they indicate that countries with low levels of electricity access put women at a severe disadvantage in their ability to enjoy the same education as those people who have high levels of access to electricity.